Put all migration tests together, remove unnecessary codes
